Fix/dashboard sales (#456)
* add course info relation

* feat: Update query to get payments, now use the course relation and filter by payment type

* feat: Update dashboard stats card sales count

This commit updates the sales count displayed in the dashboard stats cards for credit card sales and Tezos sales. Instead of relying on the length of the payments array, it now uses separate variables to track the total count of credit card sales and Tezos sales. This change ensures accurate and up-to-date sales counts in the dashboard.

Co-authored-by: Datzel <davemg71@hotmail.com>

* feat: Update payments query to filter by teacher ID

This commit updates the payments query in the user dashboard to filter the results based on the teacher's ID. Instead of using the `$_id` variable, it now uses the `$id` variable to ensure accurate retrieval of payments for the specific teacher. This change improves the accuracy of the displayed payment information in the user dashboard.

* feat: Update payments query to include credit card and Tezos sales count

This commit updates the payments query in the user dashboard to include separate variables for tracking the count of credit card sales and Tezos sales. This change ensures accurate and up-to-date sales counts in the dashboard.

Co-authored-by: Datzel <davemg71@hotmail.com>

---------

Co-authored-by: Datzel <davemg71@hotmail.com>