Add country tax validation and calculation in Payments class (#465)
* feat: Update logging level for stripe verify response

* add dependency to check european countries

* feat: Add country tax validation in Payments class

This commit adds a new method, `assertCountryTaxValidation`, to the `Payments` class. This method validates the country tax by checking if the country is a member of the European Economic Area (EEA) using the `is-european` library. The purpose of this change is to ensure that the country tax is valid before proceeding with further calculations or operations.

* update docs

* feat: Add tax calculation validation in Payments class

* feat: Refactor country tax validation in Payments class

This commit refactors the `assertCountryTaxValidation` method in the `Payments` class to improve the validation of country tax. The method now throws a `BadRequest` error if the country is not part of the European Economic Area (EEA). This change ensures that only valid countries are accepted for payment processing.