Enhance course recommendations feature and update courseRecommendation component (#466)
* feat: Add "farcaster" to cSpell dictionary

* feat: Enhance course recommendations table in the database

This commit adds a new SQL migration file to enhance the course_recommendations table in the database. It renames the "social" column to "twitter" and adds a new column "farcaster" of type TEXT. The "twitter" column is also modified to allow NULL values. Additionally, it inserts a new row into the "course_recommendations" table with the corresponding values for the columns. This change improves the structure and functionality of the course recommendations feature.

Co-authored-by: Datzel <davemg71@hotmail.com>

* refactor: Update courseRecommendation component to use "twitter" instead of "social"

This commit updates the courseRecommendation component to use the "twitter" prop instead of the "social" prop. The "twitter" prop is modified to allow NULL values and has a default value of an empty string. This change improves the clarity and consistency of the component's code.

Co-authored-by: Datzel <davemg71@hotmail.com>

* refactor: Update courseRecommendation component to use "twitter" instead of "social"

* refactor: Update courseRecommendation component to use "twitter" and "farcaster" props

* refactor: Update courseRecommendation component to include "farcaster" prop

* refactor: Add "vuex" to list of dependencies in .vscode/settings.json

* refactor: Update courseRecommendation component to include dynamic alt text

* wrapper added to apply the justify between

---------

Co-authored-by: Datzel <davemg71@hotmail.com>