Erick/feat/ilithya_requests (#484)
* Add is_countable column to modules table and update necessary records

Refactor the `modules` table in the database by adding a new column `is_countable` with a default value of `true`. Also, update specific modules to be not countable.

* Refactor module count display in buyCourse page

* Refactor module count display in buyCourse page

* feat: add subtitle prop to courseCurriculum component

* feat: Add subtitles to countable modules in public.modules table

This commit adds a new column "subtitle" to the "public.modules" table in the database. The "subtitle" column is used to store the subtitle for countable modules. The commit also updates the subtitles for specific countable modules using their IDs.

Co-authored-by: [Author Name]

* update chapter names

* update chapter names